Day4 optimization
no idea why, but I wanted to optimize day4 a bit. for very large data (240mb input file) the time went from 13seconds to 8.9. iIwas actually expecting more. but getting rid of many of the IFs in the main loop helped. probably the memory layout is not perfect, getting rid of this vector<string> and use simply a char[][] could help, and would also improve the addBorder createNewString madness.