day6 (part2)
this was kind of strange, instead of optimize the parsing fct i had to write a second one, which is also not too clean.
but the problem was kind of hard enough already, so that i was happy with the solution that is working fast enough and correctly, and i did not bother to optimize it more (we do create too many temp objects) or try to unify it with part1 parsing.